employee-recognition-api/requirements.txt
# requirements.txt
flask==0.12
pymysql
flask-sqlalchemy
GoogleAppEngineCloudStorageClient==1.9.22.1

# Only for Unit Tests
behave
requests
